+++ frontend/src/features/display/chessboard.tsx
